提交关于点赞和评论之后,异步通知的代码，未完待续...

+++ src/main/java/com/jd/wego/controller/ArticleController.java
@@ -3,7 +3,6 @@ package com.jd.wego.controller;
 import com.jd.wego.entity.Article;
 import com.jd.wego.entity.User;
 import com.jd.wego.redis.JedisService;
-import com.jd.wego.redis.UserTokenKey;
 import com.jd.wego.service.ArticleService;
 import com.jd.wego.utils.CodeMsg;
 import com.jd.wego.utils.Result;
@@ -12,9 +11,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import java.util.*;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author hbquan
@@ -30,36 +29,52 @@ public class ArticleController {
     @Autowired
     JedisService jedisService;
 
+    @Autowired
+    LoginController loginController;
+
     private static final Logger log = LoggerFactory.getLogger(ArticleController.class);
 
     @PostMapping("/insert")
     @ResponseBody
     public Result<Boolean> insertArticle(HttpServletRequest request, @RequestBody Article article){
-        LoginController loginController = new LoginController();
-        String token = loginController.getUserToken(request, LoginController.USER_TOKEN);
-        User user = jedisService.getKey(UserTokenKey.userTokenKey, token, User.class);
+        User user = loginController.getUserInfo(request);
         if(user == null){
-            System.out.println("用户未登录.......");
-            return Result.error(CodeMsg.ERROR);
+            log.info("用户未登录");
+            return Result.error(CodeMsg.NOT_LOGIN);
         }
-        log.info(user.toString());
         articleService.insertArticle(article);
         return Result.success(true);
     }
 
     @PostMapping("/edit")
     @ResponseBody
-    public Result<Boolean> editArticle(Article article){
+    public Result<Boolean> editArticle(HttpServletRequest request, @RequestBody Article article){
+        User user = loginController.getUserInfo(request);
+        if(user == null){
+            log.info("用户未登录");
+            return Result.error(CodeMsg.NOT_LOGIN);
+        }
         articleService.updateArticle(article);
         return Result.success(true);
     }
 
     @PostMapping("/delete")
     @ResponseBody
-    public Result<Boolean> deleteArticle(int articleId){
+    public Result<Boolean> deleteArticle(HttpServletRequest request, int articleId){
+        User user = loginController.getUserInfo(request);
+        if(user == null){
+            log.info("用户未登录");
+            return Result.error(CodeMsg.NOT_LOGIN);
+        }
         articleService.deleteArticle(articleId);
         return Result.success(true);
     }
 
+
+    @GetMapping("/search")
+    public Result<Boolean> searchArticle(String keyword){
+        List<Article> articleList = articleService.selectArticleByKeyword(keyword);
+        return Result.success(true);
+    }
 
 }
